@@ -1,3 +1,3 @@
-class Open < ActiveRecord::Base
+class Open < ApplicationRecord
   before_save :downcase_email
 end
