@@ -1,4 +1,4 @@
-class Show < ActiveRecord::Base
+class Show < ApplicationRecord
   extend FriendlyId
   friendly_id :name_slug_candidates, use: :slugged
 
