@@ -1,4 +1,4 @@
-class AddDirections < ActiveRecord::Migration
+class AddDirections < ActiveRecord::Migration[4.2]
   def change
     add_column :venues, :directions, :text
     add_column :venues, :contact_info, :text
