@@ -1,4 +1,4 @@
-class AddRSVPIp < ActiveRecord::Migration
+class AddRSVPIp < ActiveRecord::Migration[4.2]
   def change
     add_column :rsvps, :ip_address, :string, after: :response
   end
